requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
rembg==2.0.50
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
numpy==1.24.4
aiofiles==23.2.1
python-magic==0.4.27
pydantic==2.5.0
jinja2==3.1.2
aiohttp==3.8.5